ci: add GitHub Actions workflow for lint, test, and build
Add CI workflow that runs on push/PR to main with checkout, Node 22
setup, npm ci, lint, test, and build steps. Also fix issues that
prevented CI from passing:

- Add eslint.config.mjs (was missing, causing interactive prompt)
- Switch lint script from "next lint" to "eslint ." (Next.js 15 compat)
- Extract helper functions from route files to avoid invalid Next.js
  route exports (parseTmuxList etc. to lib/sessionHelpers, RecentPR
  interface to types/prs)
- Fix ThemeToggle to use useSyncExternalStore instead of
  useState+useEffect for hydration (lint error)
- Add null safety for sessions data in AgentStatusCards
- Fix test mocks to handle /api/sessions endpoint

Closes #25
