+++ eslint.config.mjs
@@ -1,0 +1,10 @@
+import nextConfig from "eslint-config-next";
+
+const eslintConfig = [
+  ...nextConfig,
+  {
+    ignores: [".next/"],
+  },
+];
+
+export default eslintConfig;
